examples/python/requirements.txt
# Azure SDK packages
azure-identity>=1.12.0
azure-mgmt-resource>=22.0.0
azure-mgmt-storage>=21.0.0
azure-mgmt-keyvault>=10.2.0
azure-mgmt-monitor>=5.0.0

# CLI and automation
azure-cli>=2.45.0

# Utilities
requests>=2.28.0
pyyaml>=6.0
python-dotenv>=0.19.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991